% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{Crop_image_to_each_Shapefile_polygon}
\alias{Crop_image_to_each_Shapefile_polygon}
\title{Crop image to each polygon in a shapefile}
\usage{
Crop_image_to_each_Shapefile_polygon(shapefile.dsn, shapefile.layer, image.path,
  cores, output.dir)
}
\value{
writes rasters
}
\description{
Uses each polygon in a shapefile to crop a raster.
}
\details{
writes rasters to output dir
}
\author{
Tedward Erker
}

