% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{fromUTC}
\alias{fromUTC}
\title{fromUTC}
\usage{
fromUTC(utctime)
}
\arguments{
\item{utctime}{POSIXct UTC time}
}
\value{
localtime POSIXct local time
}
\description{
Returns local time from UTC time
}
\examples{
 fromUTC( as.POSIXct("1991-06-29 05:27:25", tz = "UTC") )

}
