% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{\%h+\%}
\alias{\%h+\%}
\title{add hours}
\usage{
h \%h+\% datetime
}
\arguments{
\item{h}{number of hours}

\item{datetime}{POSIXct}
}
\value{
POSIXct and h added
}
\description{
This function adds h hours to a POSIXct
}
