% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{hoursUTC}
\alias{hoursUTC}
\title{hoursUTC}
\usage{
hoursUTC()
}
\value{
number of hours
}
\description{
This function returns time difference between UTC and Local
}
\examples{
 hoursUTC()
 #return integer equal to diference between local and utc time

}
