% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{toUTC}
\alias{toUTC}
\title{toUTC}
\usage{
toUTC(localtime)
}
\arguments{
\item{localtime}{POSIXct local time}
}
\value{
utctime POSIXct UTC time
}
\description{
Returns UTC time from local time
}
