% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{hoursUTC}
\alias{hoursUTC}
\title{hoursUTC}
\usage{
hoursUTC(to_test_time = Sys.time(), timeformat = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{to_test_time}{only for testing UTC Sys.time()}

\item{timeformat}{string format of local time, by default it will be "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\value{
number of hours
}
\description{
This function returns time difference between UTC and Local
}
