% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{toUTC}
\alias{toUTC}
\title{toUTC}
\usage{
toUTC(localtime, timeformat = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{localtime}{POSIXct local time}

\item{timeformat}{string format of local time, by default it will be "\%Y-\%m-\%d \%H:\%M:\%S"}
}
\value{
utctime POSIXct UTC time
}
\description{
Returns UTC time from local time
}
